+++ packages/eslint/src/presets/e18e.ts
@@ -1,7 +1,13 @@
-import type { TypedFlatConfigItem } from '../types'
+import type { ConfigItemRules } from '../types'
 
-export const e18eRules = {
+/**
+ * - rule: `e18e/*`
+ * - plugin: `@e18e/eslint-plugin`
+ * @see https://github.com/e18e/eslint-plugin
+ */
+export const E18E_RULES = {
+  // https://github.com/es-tooling/eslint-plugin-depend/blob/main/docs/rules/ban-dependencies.md
   'e18e/ban-dependencies': 'off',
   'e18e/prefer-array-to-sorted': 'off',
   'e18e/prefer-static-regex': 'off',
-} satisfies NonNullable<TypedFlatConfigItem['rules']>
+} satisfies ConfigItemRules