+++ packages/eslint/src/presets/eslint-comments.ts
@@ -1,5 +1,11 @@
-import type { TypedFlatConfigItem } from '../types'
+import type { ConfigItemRules } from '../types'
 
-export const eslintCommentsRules = {
+/**
+ * - rule: `eslint-comments/*`
+ * - plugin: `eslint-plugin-eslint-comments`
+ * @see https://github.com/antfu/eslint-config
+ * @see https://github.com/eslint-community/eslint-plugin-eslint-comments
+ */
+export const ESLINT_COMMENTS_RULES = {
   'eslint-comments/no-unlimited-disable': 'off',
-} satisfies NonNullable<TypedFlatConfigItem['rules']>
+} satisfies ConfigItemRules