+++ packages/eslint/src/presets/pnpm.ts
@@ -1,5 +1,11 @@
-import type { TypedFlatConfigItem } from '../types'
+import type { ConfigItemRules } from '../types'
 
-export const pnpmRules = {
+/**
+ * - rule: `pnpm/*`
+ * - plugin: `eslint-plugin-pnpm`
+ * @see https://github.com/antfu/eslint-config
+ * @see https://github.com/antfu/pnpm-workspace-utils/tree/main/packages/eslint-plugin-pnpm
+ */
+export const PNPM_RULES = {
   'pnpm/yaml-enforce-settings': 'off',
-} satisfies NonNullable<TypedFlatConfigItem['rules']>
+} satisfies ConfigItemRules