+++ packages/eslint/src/presets/react.ts
@@ -1,5 +1,10 @@
-import type { TypedFlatConfigItem } from '../types'
+import type { ConfigItemRules } from '../types'
 
-export const reactRules = {
+/**
+ * - rule: `react/*`、`react-hooks-extra/*`
+ * - plugin: `eslint-plugin-react-hooks-extra`
+ * @see https://www.npmjs.com/package/eslint-plugin-react-hooks-extra
+ */
+export const REACT_RULES = {
   'react-hooks-extra/no-direct-set-state-in-use-effect': 'off',
-} satisfies NonNullable<TypedFlatConfigItem['rules']>
+} satisfies ConfigItemRules