+++ packages/eslint/src/presets/test.ts
@@ -1,5 +1,13 @@
-import type { TypedFlatConfigItem } from '../types'
+import type { ConfigItemRules } from '../types'
 
-export const testRules = {
+/**
+ * - rule: `test/*`
+ * - original rule: `vitest/*`, `no-only-tests/*`
+ * - plugin: `@vitest/eslint-plugin`, `eslint-plugin-no-only-tests`
+ * @see https://github.com/antfu/eslint-config#test
+ * @see https://github.com/vitest-dev/eslint-plugin-vitest
+ * @see https://github.com/levibuzolic/eslint-plugin-no-only-tests
+ */
+export const TEST_RULES = {
   'test/prefer-lowercase-title': 'off',
-} satisfies NonNullable<TypedFlatConfigItem['rules']>
+} satisfies ConfigItemRules