+++ packages/eslint/src/presets/vue.ts
@@ -1,19 +1,21 @@
-import type { TypedFlatConfigItem } from '../types'
+import type { ConfigItemRules } from '../types'
 
-export const vueRules = {
+export const VUE_RULES = {
+  // Delegated to Prettier
   'vue/singleline-html-element-content-newline': 'off',
   'vue/html-closing-bracket-newline': 'off',
   'vue/html-indent': 'off',
+
   'vue/html-self-closing': [
     'warn',
     {
       html: {
-        void: 'always',
-        normal: 'always',
+        void: 'always', // Keep consistent with Prettier
+        normal: 'always', // Follow Vue Style Guide
         component: 'always',
       },
       svg: 'always',
       math: 'always',
     },
   ],
-} satisfies NonNullable<TypedFlatConfigItem['rules']>
+} satisfies ConfigItemRules