+++ packages/eslint/src/types/define-config.ts
@@ -3,5 +3,6 @@ import type { OptionsConfig, TypedFlatConfigItem } from '@antfu/eslint-config'
 
 export type AntfuOptions = OptionsConfig & Omit<TypedFlatConfigItem, 'files'>
 export type Composer = ReturnType<typeof antfu>
+export type ConfigItemRules = NonNullable<TypedFlatConfigItem['rules']>
 export type { TypedFlatConfigItem }
 export type UserFlatConfig = Parameters<typeof antfu>[1]